requirements.txt
discord.py
Flask
Pillow
opencv-python
sounddevice
numpy
psutil
pynput
pyaudiowpatch
